+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 **Retail:**  
 - Added BetterCooldownManager import to installer and config  
 - Added Blizzard Edit Mode import guide to installer and config  
+- Added :status versions of the luckyone friendly abbrev name Tags  
 - Adjusted profile defaults for new ElvUI overAbsorb code  
 - Updated BigWigs profiles (Bar fixes & Disabled Private Aura anchor)  
 - Updated ElvUI Nameplates (Friendly players name-only and class colored in instances)